+++ src/components/common/index.ts
@@ -3,7 +3,6 @@ export { PaginationList } from './pagination/PaginationList';
 export { PaginationSelect } from './pagination/PaginationSelect';
 
 export { SampleList } from './SampleList';
-export { SampleDetail } from './SampleDetail';
 export { SearchBar } from './SearchBar';
 export { SubTitle } from './SubTitle';
 export { UrlTabs, type TabItem } from './UrlTabs';
