+++ src/components/product/index.ts
@@ -1,9 +1,9 @@
-export { CarpetTab } from './CarpetTab';
-export { CarpetTilesTab } from './CarpetTilesTab';
-export { DecoTab } from './DecoTab';
-export { DeluxeTab } from './DeluxeTab';
-export { SelfTab } from './SelfTab';
-export { WoodTab } from './WoodTab';
+export { CarpetTab } from './tabs/CarpetTab';
+export { CarpetTilesTab } from './tabs/CarpetTilesTab';
+export { DecoTab } from './tabs/DecoTab';
+export { DeluxeTab } from './tabs/DeluxeTab';
+export { SelfTab } from './tabs/SelfTab';
+export { WoodTab } from './tabs/WoodTab';
 
 export { InquiryDialog } from './InquiryDialog';
 export { ProductCounter } from './ProductCounter';
