+++ src/components/product/tabs/CarpetTab.tsx
