+++ src/components/product/tabs/CarpetTilesTab.tsx
