+++ src/components/product/tabs/DecoTab.tsx
