+++ src/components/product/tabs/DeluxeTab.tsx
