+++ src/components/product/tabs/SelfTab.tsx
