+++ src/components/product/tabs/WoodTab.tsx
